Move follow controls to profiles
